# docs/student-linking-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Edvanz — Student↔Teacher Linking (Request/Approval)",
    "version": "1.0",
...
  },
  "servers": [
    {
      "url": "https://app-edvanz-api-prod.azurewebsites.net",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/studentuser/me": {
      "get": {
        "tags": ["Student — profile & dashboard"],
        "summary": "My student profile",
        "description": "Returns the authenticated student's profile: account code (used by PARENTS to link to this child), language, status, first-login flag and the count of Active linked teachers.",
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentProfileEnvelope" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/StudentNotResolved" }
        }
      }
    },
    "/api/studentuser/me/profile": {
      "put": {
        "tags": ["Student — profile & dashboard"],
        "summary": "Update my profile settings",
        "description": "Currently supports the UI language preference only (name/phone/password are User-module operations).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateStudentProfileRequest" },
              "examples": {
                "arabic": { "value": { "languagePreference": "ar" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentProfileEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/StudentNotResolved" }
        }
      }
    },
    "/api/studentuser/me/dashboard": {
      "get": {
        "tags": ["Student — profile & dashboard"],
        "summary": "My dashboard",
        "description": "First-login flag (drives the empty-state \"Add Teacher\" prompt), the student's account code, and the teacher entries — one per teacher, latest state, same shape as `GET /me/teachers`.",
        "responses": {
          "200": {
            "description": "Dashboard",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentDashboardEnvelope" },
                "examples": {
                  "withPendingAndActive": {
                    "value": {
                      "success": true,
                      "code": "Success",
                      "message": "Done successfully",
                      "data": {
                        "isFirstLogin": false,
                        "studentAccountCode": "STU4X7B2KM",
                        "linkedTeachers": [
                          {
                            "linkId": 12,
                            "status": "Pending",
                            "requestedAt": "2026-07-12T18:40:00Z",
                            "respondedAt": null,
                            "teacherCode": "48291057",
                            "teacherFullName": "Ahmed Mohamed",
                            "subjectName": "Mathematics",
                            "linkedAt": "2026-07-12T18:40:00Z",
                            "isEnrollmentActive": false,
                            "visibilityAttendance": true,
                            "visibilityPayment": true,
                            "visibilityHomework": true,
                            "visibilityExamDefault": false
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/StudentNotResolved" }
        }
      }
    },
    "/api/studentuser/me/teachers": {
      "get": {
        "tags": ["Student — teacher linking"],
        "summary": "My teachers — with request status",
        "description": "One entry per teacher (the LATEST link row), so the student sees the current state: `Pending` (waiting for the teacher), `Active` (accepted — modules unlocked per the visibility flags), `Rejected`, `Unlinked`, `RemovedByTeacher`, `CancelledByStudent`. Live entries (Pending, Active) are sorted first. **This endpoint is how the app shows whether a request was accepted or rejected** (a push notification is also sent).",
        "responses": {
          "200": {
            "description": "Teacher entries",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentTeacherListEnvelope" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/StudentNotResolved" }
        }
      }
    },
    "/api/studentuser/me/link-requests": {
      "post": {
        "tags": ["Student — teacher linking"],
        "summary": "Send a link request to a teacher",
        "description": "Creates a `Pending` request. `teacherCode` is the teacher's public 8-digit code; `studentName` is typed by the student so the teacher can recognize them; `studentCode` is OPTIONAL — when it matches a roster record it pre-suggests the binding on the teacher's accept screen. The link only becomes usable after the teacher accepts.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateLinkRequestRequest" },
              "examples": {
                "withCode": {
                  "summary": "Student knows their code at this teacher",
                  "value": { "teacherCode": "48291057", "studentName": "Mohamed Ahmed", "studentCode": "A5" }
                },
                "withoutCode": {
                  "summary": "No code — teacher will pick the roster record manually",
                  "value": { "teacherCode": "48291057", "studentName": "Mohamed Ahmed" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request created (entry returned with `status: Pending`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentTeacherEntryEnvelope" }
              }
            }
          },
          "400": {
            "description": "Invalid teacher code format, or missing student name (`InvalidTeacherCode`, `StudentNameRequired`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "No active teacher with this code (`TeacherNotFound`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          },
          "409": {
            "description": "Already linked (`TeacherAlreadyLinked`) or a request is already pending (`LinkRequestAlreadyPending`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          }
        }
      }
    },
    "/api/studentuser/me/teachers/{teacherId}": {
      "delete": {
        "tags": ["Student — teacher linking"],
        "summary": "Remove a teacher (cancel pending / unlink active)",
        "description": "Cancels a still-`Pending` request (state becomes `CancelledByStudent`, code `LinkRequestCancelled`) or unlinks an `Active` link (state `Unlinked`, code `TeacherUnlinkSuccess`). History is preserved; a new request can be sent later.",
        "parameters": [
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" },
            "description": "The teacher's id (from the entry's teacher data)."
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled or unlinked (`data: true`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BooleanEnvelope" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "No pending request or active link with this teacher (`LinkNotFound`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          }
        }
      }
    },
    "/api/studentuser/by-code/{accountCode}": {
      "get": {
        "tags": ["Student — profile & dashboard"],
        "summary": "Look up a student by account code (parent linking)",
        "description": "Used by the PARENT app (Method A child linking). Any authenticated role may call it. Returns minimal profile info.",
        "parameters": [
          {
            "name": "accountCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "STU4X7B2KM"
          }
        ],
        "responses": {
          "200": {
            "description": "Basic student profile",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StudentProfileEnvelope" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": {
            "description": "No active student with this code (`StudentUserNotFound`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "JWT missing or expired",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
...
      "BadRequest": {
        "description": "Validation failure",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "StudentNotResolved": {
        "description": "The authenticated user has no student account",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
...
    },
    "schemas": {
      "LinkStatus": {
        "type": "string",
        "enum": ["Pending", "Active", "Rejected", "Unlinked", "RemovedByTeacher", "CancelledByStudent"],
        "description": "Pending = waiting for the teacher; Active = accepted (modules unlocked); Rejected = teacher declined; Unlinked = student removed an active link; RemovedByTeacher = teacher removed the student; CancelledByStudent = student cancelled their own pending request."
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "code": { "type": "string", "description": "Stable, language-independent key to branch on (e.g. LinkRequestAlreadyPending)" },
          "message": { "type": "string", "description": "Localized (Accept-Language: en | ar)" }
        }
      },
      "BooleanEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "data": { "type": "boolean", "example": true }
        }
      },
      "StudentProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "userId": { "type": "integer", "format": "int64" },
          "studentAccountCode": { "type": "string", "example": "STU4X7B2KM" },
          "fullName": { "type": "string" },
          "email": { "type": "string", "nullable": true },
          "phoneNumber": { "type": "string", "nullable": true },
          "languagePreference": { "type": "string", "nullable": true, "example": "ar" },
          "accountStatus": { "type": "string", "example": "Active" },
          "isFirstLogin": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "linkedTeacherCount": { "type": "integer", "description": "Active links only" }
        }
      },
      "StudentProfileEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/StudentProfile" }
        }
      },
      "UpdateStudentProfileRequest": {
        "type": "object",
        "properties": {
          "languagePreference": { "type": "string", "enum": ["en", "ar"], "nullable": true }
        }
      },
      "StudentTeacherEntry": {
        "type": "object",
        "description": "One teacher on the student's list — the LATEST link row for that teacher.",
        "properties": {
          "linkId": { "type": "integer", "format": "int64" },
          "status": { "$ref": "#/components/schemas/LinkStatus" },
          "requestedAt": { "type": "string", "format": "date-time", "nullable": true, "description": "When the student sent the request" },
          "respondedAt": { "type": "string", "format": "date-time", "nullable": true, "description": "When the teacher accepted/rejected" },
          "teacherCode": { "type": "string", "example": "48291057" },
          "teacherFullName": { "type": "string" },
          "subjectName": { "type": "string", "description": "Localized to the student's language preference" },
          "linkedAt": { "type": "string", "format": "date-time", "description": "When the link became Active (mirrors requestedAt while still pending)" },
          "isEnrollmentActive": { "type": "boolean", "description": "True only when status=Active AND the teacher's roster record still exists" },
          "visibilityAttendance": { "type": "boolean" },
          "visibilityPayment": { "type": "boolean" },
          "visibilityHomework": { "type": "boolean" },
          "visibilityExamDefault": { "type": "boolean" }
        }
      },
      "StudentTeacherEntryEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string", "example": "LinkRequestCreated" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/StudentTeacherEntry" }
        }
      },
      "StudentTeacherListEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/StudentTeacherEntry" }
          }
        }
      },
      "StudentDashboard": {
        "type": "object",
        "properties": {
          "isFirstLogin": { "type": "boolean", "description": "True until the student sends their first link request — drives the empty-state prompt" },
          "studentAccountCode": { "type": "string" },
          "linkedTeachers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/StudentTeacherEntry" }
          }
        }
      },
      "StudentDashboardEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/StudentDashboard" }
        }
      },
      "CreateLinkRequestRequest": {
        "type": "object",
        "required": ["teacherCode", "studentName"],
        "properties": {
          "teacherCode": { "type": "string", "description": "The teacher's public 8-digit code", "example": "48291057" },
          "studentName": { "type": "string", "maxLength": 200, "description": "Typed by the student — shown verbatim to the teacher" },
          "studentCode": { "type": "string", "maxLength": 20, "nullable": true, "description": "OPTIONAL — the code this teacher assigned the student (e.g. A5); powers the accept-screen suggestion" }
        }
      },
...
    }
  }
}